split server.rs
