+++ examples/server.rs
@@ -7,7 +7,7 @@ use structopt::StructOpt;
 use tokio::{fs::File, io::AsyncReadExt};
 use tracing::{error, info, trace_span};
 
-use h3::{error::ErrorLevel, quic::BidiStream, server::server::RequestStream};
+use h3::{error::ErrorLevel, quic::BidiStream, server::stream::RequestStream};
 use h3_quinn::quinn;
 
 #[derive(StructOpt, Debug)]
@@ -116,7 +116,7 @@ async fn main() -> Result<(), Box<dyn std::error::Error>> {
                     info!("new connection established");
 
                     let mut h3_conn =
-                        h3::server::server::Connection::new(h3_quinn::Connection::new(conn))
+                        h3::server::connection::Connection::new(h3_quinn::Connection::new(conn))
                             .await
                             .unwrap();
